+++ solver/src/main/scala/com/htmlism/spawningpool/combinatorics/specialized/HomogenousCombinator.scala
@@ -1,0 +1,11 @@
+package com.htmlism.spawningpool.combinatorics.specialized
+
+import scala.reflect.ClassTag
+
+import com.htmlism.spawningpool.combinatorics
+
+trait HomogenousCombinator[A] extends combinatorics.HomogenousCombinator[A] {
+  implicit def classTag: ClassTag[A]
+
+  override def fill(size: Int): (=> A) => Seq[A] = { fill => Array.fill[A](size)(fill) }
+}